Image Classification Model
This model is used to distinguish if an uploaded image by the user is a flood or a landslide image. The dataset I used for this is in https://drive.google.com/drive/folders/1FwuNRtJl9czXwvFfFlJW3Hfvy0UxTb-o?usp=sharing. 

Make sure to change the PATH NAME Before running and path is in the format of "/ " instead of " \"